Prove in tests that Jetty can serve JSP with custom tags
