+++ src/main/resources/webroot/WEB-INF/taglib/bye.tld
@@ -5,10 +5,11 @@
 <taglib>
     <tlib-version>1.0</tlib-version>
     <jsp-version>1.2</jsp-version>
-    <short-name>Leaving a lot behind</short-name>
+    <short-name>tags</short-name>
 
     <tag>
         <name>scripting</name>
         <tag-class>org.eclipse.jetty.demo.tags.Leaving</tag-class>
+        <body-content>empty</body-content>
     </tag>
 </taglib>
