+++ src/main/resources/webroot/test/tagclass.jsp
@@ -1,0 +1,7 @@
+<%@ taglib uri="/WEB-INF/taglib/bye.tld" prefix="leave" %>
+
+<html>
+<body>
+    <leave:scripting />
+</body>
+</html>
